feat: add node flush
