+++ pkg/runtime/node.go
@@ -2,7 +2,6 @@ package runtime
 
 import (
 	"context"
-	"encoding/json"
 	"sync"
 	"time"
 
@@ -17,7 +16,6 @@ import (
 	"github.com/tkeel-io/core/pkg/placement"
 	"github.com/tkeel-io/core/pkg/repository"
 	"github.com/tkeel-io/core/pkg/repository/dao"
-	"github.com/tkeel-io/core/pkg/resource/tseries"
 	"github.com/tkeel-io/core/pkg/types"
 	"github.com/tkeel-io/core/pkg/util"
 	xkafka "github.com/tkeel-io/core/pkg/util/kafka"
@@ -206,122 +204,6 @@ func (n *Node) getGlobalData(en Entity) (res []byte) {
 	}
 	return globalData.GetRaw()
 }
-
-func (n *Node) FlushEntity(ctx context.Context, en Entity) error {
-	log.L().Debug("flush entity", zfield.Eid(en.ID()), zfield.Value(string(en.Raw())))
-
-	// 1. flush state.
-	if err := n.resourceManager.Repo().PutEntity(ctx, en.ID(), en.Raw()); nil != err {
-		log.L().Error("flush entity state storage", zap.Error(err), zfield.Eid(en.ID()))
-		return errors.Wrap(err, "flush entity into state storage")
-	}
-
-	// 2. flush search engine data.
-	// 2.1 flush search global data.
-	globalData := n.getGlobalData(en)
-	if _, err := n.resourceManager.Search().IndexBytes(ctx, en.ID(), globalData); nil != err {
-		log.L().Error("flush entity search engine", zap.Error(err), zfield.Eid(en.ID()))
-		//			return errors.Wrap(err, "flush entity into search engine")
-	}
-
-	// 2.2 flush search model data.
-	// TODO.
-
-	// 3. flush timeseries data.
-
-	en.Properties()
-	if err := n.flushTimeSeries(ctx, en); nil != err {
-		log.L().Error("flush entity timeseries database", zap.Error(err), zfield.Eid(en.ID()))
-	}
-	return nil
-}
-
-func (n *Node) flushTimeSeries(ctx context.Context, en Entity) (err error) {
-	tsData := en.GetProp("telemetry")
-	var flushData []*tseries.TSeriesData
-	log.Info("tsData: ", tsData)
-	var res interface{}
-
-	err = json.Unmarshal(tsData.Raw(), &res)
-	if nil != err {
-		log.L().Warn("parse json type", zap.Error(err))
-		return
-	}
-	tss, ok := res.(map[string]interface{})
-	if ok {
-		for k, v := range tss {
-			switch tsOne := v.(type) {
-			case map[string]interface{}:
-				if ts, ok := tsOne["ts"]; ok {
-					tsItem := tseries.TSeriesData{
-						Measurement: "keel",
-						Tags:        map[string]string{"id": en.ID()},
-						Fields:      map[string]float32{},
-						Timestamp:   0,
-					}
-					switch tttV := tsOne["value"].(type) {
-					case float64:
-						tsItem.Fields[k] = float32(tttV)
-						timestamp, _ := ts.(float64)
-						tsItem.Timestamp = int64(timestamp) * 1e6
-						flushData = append(flushData, &tsItem)
-					case float32:
-						tsItem.Fields[k] = tttV
-						timestamp, _ := ts.(float64)
-						tsItem.Timestamp = int64(timestamp) * 1e6
-						flushData = append(flushData, &tsItem)
-					}
-					continue
-				}
-			default:
-				log.Info(tsOne)
-			}
-		}
-	}
-	_, err = n.resourceManager.TSDB().Write(ctx, &tseries.TSeriesRequest{
-		Data:     flushData,
-		Metadata: map[string]string{},
-	})
-	return errors.Wrap(err, "write ts db error")
-}
-
-func (n *Node) RemoveEntity(ctx context.Context, en Entity) error {
-	var err error
-
-	// recover entity state.
-	defer func() {
-		if nil != err {
-			if innerErr := n.FlushEntity(ctx, en); nil != innerErr {
-				log.L().Error("remove entity failed, recover entity state failed", zfield.Eid(en.ID()),
-					zfield.Reason(err.Error()), zap.Error(innerErr), zfield.Value(string(en.Raw())))
-			}
-		}
-	}()
-
-	// 1. 从状态存储中删除（可标记）
-	if err := n.resourceManager.Repo().
-		DelEntity(ctx, en.ID()); nil != err {
-		log.L().Error("remove entity from state storage",
-			zap.Error(err), zfield.Eid(en.ID()), zfield.Value(string(en.Raw())))
-		return errors.Wrap(err, "remove entity from state storage")
-	}
-
-	// 2. 从搜索中删除（可标记）
-	if _, err := n.resourceManager.Search().
-		DeleteByID(ctx, &v1.DeleteByIDRequest{
-			Id:     en.ID(),
-			Owner:  en.Owner(),
-			Source: en.Source(),
-		}); nil != err {
-		log.L().Error("remove entity from state search engine",
-			zap.Error(err), zfield.Eid(en.ID()), zfield.Value(string(en.Raw())))
-		return errors.Wrap(err, "remove entity from state search engine")
-	}
-
-	// 3. 删除实体相关的 Expression.
-	return nil
-}
-
 func parseExpression(expr repository.Expression, version int) (map[string]*ExpressionInfo, error) {
 	exprIns, err := expression.NewExpr(expr.Expression, nil)
 	if nil != err {
